Close-to-perfect recreation of AUDIO_WF.LIB v1.09, in terms of the (eventually) generated code. A minor exception is one register read in FX_SetupCard, although the behaviors are equivalent (this also applies to the mostly complete 1.04 recreation).
Additionally:
- Took the chance to change '#include "interrup.h"' into
'#include "interrupt.h"' anywhere, for 1.04, 1.09 and 1.1.
- Similarly added '#include "memcheck.h"' for the few sources in
which it's mentioned in the original OBJ (LIB) files, presumably
due to the presence of additional debugging information. It was
possibly included in *all* C files, but this can be revisited later.
- Interestingly, it looks like _MIDI_ServiceRoutine is public in 1.09,
but *not* in 1.04 or 1.1
- "Programmed by Jim Dose" string (FX_MAN.C, MUSIC.C): For some weird
reason the character of \xE9 was used in a preceding restoration
effort, instead of \x82.
- It's really FX_MAN.C which had most changes, generally re-using
restoration efforts for 1.04, with a few exceptions (mostly for GUS).
A great deal of the diffs is related to PCFX and ADLIBFX
specific code relevant for 1.04 and 1.09, but there are
also GUS-related code pieces which apply to 1.09 and 1.1
(if not changed in-between in other means).